Cleanup argv
